# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FitTrack API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ExerciseCategory = typeof ExerciseCategory[keyof typeof ExerciseCategory];


export const ExerciseCategory = {
  strength: 'strength',
  cardio: 'cardio',
  flexibility: 'flexibility',
  bodyweight: 'bodyweight',
} as const;

export type ExerciseMuscleGroup = typeof ExerciseMuscleGroup[keyof typeof ExerciseMuscleGroup];


export const ExerciseMuscleGroup = {
  chest: 'chest',
  back: 'back',
  shoulders: 'shoulders',
  arms: 'arms',
  core: 'core',
  legs: 'legs',
  fullBody: 'fullBody',
  cardio: 'cardio',
} as const;

export interface Exercise {
  id: number;
  name: string;
  category: ExerciseCategory;
  muscleGroup: ExerciseMuscleGroup;
  /** @nullable */
  description?: string | null;
}

export type ExerciseInputCategory = typeof ExerciseInputCategory[keyof typeof ExerciseInputCategory];


export const ExerciseInputCategory = {
  strength: 'strength',
  cardio: 'cardio',
  flexibility: 'flexibility',
  bodyweight: 'bodyweight',
} as const;

export type ExerciseInputMuscleGroup = typeof ExerciseInputMuscleGroup[keyof typeof ExerciseInputMuscleGroup];


export const ExerciseInputMuscleGroup = {
  chest: 'chest',
  back: 'back',
  shoulders: 'shoulders',
  arms: 'arms',
  core: 'core',
  legs: 'legs',
  fullBody: 'fullBody',
  cardio: 'cardio',
} as const;

export interface ExerciseInput {
  /** @minLength 1 */
  name: string;
  category: ExerciseInputCategory;
  muscleGroup: ExerciseInputMuscleGroup;
  description?: string;
}

export type ExerciseUpdateCategory = typeof ExerciseUpdateCategory[keyof typeof ExerciseUpdateCategory];


export const ExerciseUpdateCategory = {
  strength: 'strength',
  cardio: 'cardio',
  flexibility: 'flexibility',
  bodyweight: 'bodyweight',
} as const;

export type ExerciseUpdateMuscleGroup = typeof ExerciseUpdateMuscleGroup[keyof typeof ExerciseUpdateMuscleGroup];


export const ExerciseUpdateMuscleGroup = {
  chest: 'chest',
  back: 'back',
  shoulders: 'shoulders',
  arms: 'arms',
  core: 'core',
  legs: 'legs',
  fullBody: 'fullBody',
  cardio: 'cardio',
} as const;

export interface ExerciseUpdate {
  /** @minLength 1 */
  name?: string;
  category?: ExerciseUpdateCategory;
  muscleGroup?: ExerciseUpdateMuscleGroup;
  description?: string;
}

export type WorkoutStatus = typeof WorkoutStatus[keyof typeof WorkoutStatus];


export const WorkoutStatus = {
  planned: 'planned',
  completed: 'completed',
} as const;

export interface Workout {
  id: number;
  name: string;
  date: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
  status: WorkoutStatus;
  /** @nullable */
  totalVolume?: number | null;
}

export type WorkoutInputStatus = typeof WorkoutInputStatus[keyof typeof WorkoutInputStatus];


export const WorkoutInputStatus = {
  planned: 'planned',
  completed: 'completed',
} as const;

export interface WorkoutInput {
  /** @minLength 1 */
  name: string;
  date: string;
  durationMinutes?: number;
  notes?: string;
  status?: WorkoutInputStatus;
}

export type WorkoutUpdateStatus = typeof WorkoutUpdateStatus[keyof typeof WorkoutUpdateStatus];


export const WorkoutUpdateStatus = {
  planned: 'planned',
  completed: 'completed',
} as const;

export interface WorkoutUpdate {
  /** @minLength 1 */
  name?: string;
  date?: string;
  durationMinutes?: number;
  notes?: string;
  status?: WorkoutUpdateStatus;
}

export interface WorkoutExercise {
  id: number;
  workoutId: number;
  exerciseId: number;
  exerciseName: string;
  sets: number;
  /** @nullable */
  reps: number | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  notes?: string | null;
  orderIndex?: number;
}

export interface WorkoutExerciseInput {
  exerciseId: number;
  sets: number;
  reps?: number;
  weightKg?: number;
  durationSeconds?: number;
  notes?: string;
  orderIndex?: number;
}

export type WorkoutDetailStatus = typeof WorkoutDetailStatus[keyof typeof WorkoutDetailStatus];


export const WorkoutDetailStatus = {
  planned: 'planned',
  completed: 'completed',
} as const;

export interface WorkoutDetail {
  id: number;
  name: string;
  date: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
  status: WorkoutDetailStatus;
  /** @nullable */
  totalVolume?: number | null;
  exercises: WorkoutExercise[];
}

export interface StatsSummary {
  totalWorkouts: number;
  completedWorkouts: number;
  currentStreak: number;
  totalVolumeKg: number;
  thisWeekWorkouts: number;
  /** @nullable */
  avgDurationMinutes: number | null;
}

export interface WeeklyStats {
  week: string;
  workoutCount: number;
  totalVolumeKg: number;
}

export interface MuscleGroupStat {
  muscleGroup: string;
  count: number;
}