# lib/api-zod/src/generated/types/exercise.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FitTrack API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExerciseCategory } from './exerciseCategory';
import type { ExerciseMuscleGroup } from './exerciseMuscleGroup';

export interface Exercise {
  id: number;
  name: string;
  category: ExerciseCategory;
  muscleGroup: ExerciseMuscleGroup;
  /** @nullable */
  description?: string | null;
}