# lib/api-zod/src/generated/types/exerciseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FitTrack API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExerciseInputCategory } from './exerciseInputCategory';
import type { ExerciseInputMuscleGroup } from './exerciseInputMuscleGroup';

export interface ExerciseInput {
  /** @minLength 1 */
  name: string;
  category: ExerciseInputCategory;
  muscleGroup: ExerciseInputMuscleGroup;
  description?: string;
}