# lib/api-zod/src/generated/types/exerciseUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FitTrack API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExerciseUpdateCategory } from './exerciseUpdateCategory';
import type { ExerciseUpdateMuscleGroup } from './exerciseUpdateMuscleGroup';

export interface ExerciseUpdate {
  /** @minLength 1 */
  name?: string;
  category?: ExerciseUpdateCategory;
  muscleGroup?: ExerciseUpdateMuscleGroup;
  description?: string;
}