# lib/api-zod/src/generated/types/muscleGroupStat.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FitTrack API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MuscleGroupStat {
  muscleGroup: string;
  count: number;
}