# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FitTrack API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalWorkouts: number;
  completedWorkouts: number;
  currentStreak: number;
  totalVolumeKg: number;
  thisWeekWorkouts: number;
  /** @nullable */
  avgDurationMinutes: number | null;
}