# lib/api-zod/src/generated/types/workout.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FitTrack API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutStatus } from './workoutStatus';

export interface Workout {
  id: number;
  name: string;
  date: Date;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
  status: WorkoutStatus;
  /** @nullable */
  totalVolume?: number | null;
}