# lib/api-zod/src/generated/types/workoutDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FitTrack API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutDetailStatus } from './workoutDetailStatus';
import type { WorkoutExercise } from './workoutExercise';

export interface WorkoutDetail {
  id: number;
  name: string;
  date: Date;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
  status: WorkoutDetailStatus;
  /** @nullable */
  totalVolume?: number | null;
  exercises: WorkoutExercise[];
}