# lib/api-zod/src/generated/types/workoutExercise.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FitTrack API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkoutExercise {
  id: number;
  workoutId: number;
  exerciseId: number;
  exerciseName: string;
  sets: number;
  /** @nullable */
  reps: number | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  notes?: string | null;
  orderIndex?: number;
}