# lib/api-zod/src/generated/types/workoutExerciseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FitTrack API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkoutExerciseInput {
  exerciseId: number;
  sets: number;
  reps?: number;
  weightKg?: number;
  durationSeconds?: number;
  notes?: string;
  orderIndex?: number;
}