# lib/api-zod/src/generated/types/workoutUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FitTrack API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutUpdateStatus } from './workoutUpdateStatus';

export interface WorkoutUpdate {
  /** @minLength 1 */
  name?: string;
  date?: Date;
  durationMinutes?: number;
  notes?: string;
  status?: WorkoutUpdateStatus;
}